refactor(RevieweeResponseDto): Nested class 추출
Related to: #65
