+++ src/main/java/com/hf/healthfriend/domain/review/dto/response/ReviewDetailPerEvaluationType.java
@@ -1,0 +1,7 @@
+package com.hf.healthfriend.domain.review.dto.response;
+
+public record ReviewDetailPerEvaluationType(
+        Integer reviewDetailId,
+        Long reviewDetailCount
+) {
+}
